+++ gl-style-editor/__init__.py
@@ -1,0 +1,1 @@
+from .glse import run
