+++ pyproject.toml
@@ -8,6 +8,7 @@ authors = [
     "Yannick Lapointe <yannicklapointe77@gmail.com>",
 ]
 readme = "README.md"
+packages = [{include = "glse.py", from = "gl-style-editor"}]
 
 [tool.poetry.dependencies]
 python = ">=3.10,<3.13"
